refactor: refactor package architecture
