+++ src/main/java/com/pencilwith/apiserver/auth/service/CustomUserDetailService.java
@@ -1,0 +1,37 @@
+package com.pencilwith.apiserver.auth.service;
+
+import com.pencilwith.apiserver.common.model.entity.User;
+import com.pencilwith.apiserver.auth.repository.UserRepository;
+import java.util.List;
+import java.util.stream.Collectors;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class CustomUserDetailService implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String userId) {
+        return userRepository.findById(userId)
+                .map(user -> createUser(user))
+                .orElseThrow(() -> new UsernameNotFoundException(userId + " -> 데이터베이스에서 찾을 수 없습니다."));
+    }
+
+    private org.springframework.security.core.userdetails.User createUser(User user) {
+        List<GrantedAuthority> grantedAuthorities = user.getUserAuthorities().stream()
+                .map(authority -> new SimpleGrantedAuthority(authority.getAuthority().getType().toString()))
+                .collect(Collectors.toList());
+
+        return new org.springframework.security.core.userdetails.User(user.getId(), user.getPassword(), grantedAuthorities);
+    }
+}
