+++ src/main/java/com/pencilwith/apiserver/common/enums/AuthorityType.java
@@ -1,0 +1,5 @@
+package com.pencilwith.apiserver.common.enums;
+
+public enum AuthorityType {
+    ROLE_USER, ROLE_ADMIN
+}
