+++ src/main/java/com/pencilwith/apiserver/common/enums/CareerType.java
@@ -1,0 +1,5 @@
+package com.pencilwith.apiserver.common.enums;
+
+public enum CareerType {
+    NEWBIE, INTERMEDIATE, SENIOR
+}
