+++ src/main/java/com/pencilwith/apiserver/common/enums/GenderType.java
@@ -1,0 +1,5 @@
+package com.pencilwith.apiserver.common.enums;
+
+public enum GenderType {
+    MALE, FEMALE
+}
