+++ src/main/java/com/pencilwith/apiserver/common/model/dto/AuthenticationDto.java
@@ -1,0 +1,19 @@
+package com.pencilwith.apiserver.common.model.dto;
+
+import lombok.Getter;
+
+@Getter
+public class AuthenticationDto {
+
+    private String userId;
+
+    private String password;
+
+    public AuthenticationDto() {
+    }
+
+    public AuthenticationDto(String userId, String password) {
+        this.userId = userId;
+        this.password = password;
+    }
+}
