+++ src/main/java/com/pencilwith/apiserver/common/model/dto/UserInfoResponseDto.java
@@ -1,0 +1,20 @@
+package com.pencilwith.apiserver.common.model.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UserInfoResponseDto {
+
+    private String loginType;
+
+    @JsonProperty("id")
+    private String userId;
+
+    private String profileImage;
+
+    public UserInfoResponseDto() {
+    }
+}
