+++ src/main/java/com/pencilwith/apiserver/config/auth/AuthConfig.java
@@ -1,0 +1,24 @@
+package com.pencilwith.apiserver.config.auth;
+
+import com.pencilwith.apiserver.common.model.dto.UserInfoResponseDto;
+import java.util.LinkedHashMap;
+import java.util.Map;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class AuthConfig {
+
+    private static final int MAX_SIZE = 1000;
+
+    // TODO: 스레드 세이프하도록 변경 필요
+    @Bean
+    public LinkedHashMap<String, UserInfoResponseDto> userInfoStorage() {
+        return new LinkedHashMap<>(MAX_SIZE * 10 / 7, 0.7f, true) {
+            @Override
+            protected boolean removeEldestEntry(Map.Entry<String, UserInfoResponseDto> eldest) {
+                return size() > MAX_SIZE;
+            }
+        };
+    }
+}
