--- src/main/java/com/pencilwith/apiserver/controller/HealthcheckController.java
@@ -1,15 +1,0 @@
-package com.pencilwith.apiserver.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/")
-public class HealthcheckController {
-
-    @GetMapping("/hello")
-    public String hello() {
-        return "hello";
-    }
-}
