+++ src/main/java/com/pencilwith/apiserver/healthcheck/HealthcheckController.java
@@ -1,0 +1,15 @@
+package com.pencilwith.apiserver.healthcheck;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/")
+public class HealthcheckController {
+
+    @GetMapping("/hello")
+    public String hello() {
+        return "hello";
+    }
+}
