--- src/main/java/com/pencilwith/apiserver/model/dto/UserInfoResponseDto.java
@@ -1,20 +1,0 @@
-package com.pencilwith.apiserver.model.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserInfoResponseDto {
-
-    private String loginType;
-
-    @JsonProperty("id")
-    private String userId;
-
-    private String profileImage;
-
-    public UserInfoResponseDto() {
-    }
-}
