--- src/main/java/com/pencilwith/apiserver/model/enums/AuthorityType.java
@@ -1,5 +1,0 @@
-package com.pencilwith.apiserver.model.enums;
-
-public enum AuthorityType {
-    ROLE_USER, ROLE_ADMIN
-}
