--- src/main/java/com/pencilwith/apiserver/model/enums/CareerType.java
@@ -1,5 +1,0 @@
-package com.pencilwith.apiserver.model.enums;
-
-public enum CareerType {
-    NEWBIE, INTERMEDIATE, SENIOR
-}
