--- src/main/java/com/pencilwith/apiserver/model/enums/GenderType.java
@@ -1,5 +1,0 @@
-package com.pencilwith.apiserver.model.enums;
-
-public enum GenderType {
-    MALE, FEMALE
-}
