--- src/main/java/com/pencilwith/apiserver/model/enums/LocationType.java
@@ -1,5 +1,0 @@
-package com.pencilwith.apiserver.model.enums;
-
-public enum LocationType {
-    SEOUL, INCHEON
-}
