--- src/main/java/com/pencilwith/apiserver/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.pencilwith.apiserver.repository;
-
-import com.pencilwith.apiserver.model.entity.User;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, String> {
-
-    Optional<User> findUserByUsername(String username);
-}
