--- src/main/java/com/pencilwith/apiserver/service/GoogleService.java
@@ -1,12 +1,0 @@
-package com.pencilwith.apiserver.service;
-
-import com.pencilwith.apiserver.model.dto.AuthenticationResultDto;
-import org.springframework.stereotype.Service;
-
-@Service
-public class GoogleService extends AuthService {
-
-    public AuthenticationResultDto processAuthentication(String authorizationCode) {
-        return null;
-    }
-}
