+++ src/main/resources/data.sql
@@ -4,6 +4,6 @@ INSERT INTO AUTHORITY (ID, TYPE) values (2, 'ROLE_ADMIN');
 INSERT INTO PUBLIC.USER (ID, BIRTH, CAREER_TYPE, GENDER_TYPE, INTRODUCTION, LOCATION_TYPE, PASSWORD, PROFILE_IMAGE, USERNAME) VALUES ('kakao1', '2021-03-30', 'SENIOR', 'MALE', '자기소개', 'SEOUL', '$2a$10$EUcWG8grzRnfWr0NnGT5ZeKKWq03ALSCgBVKXTqju0xwG1NP.FzW.', 'https://www.google.com/url?sa=i&url=https%3A%2F%2Ffirpeng.tistory.com%2F103&psig=AOvVaw2eQNj_SndJ_UcO5fS-SsSk&ust=1617196831404000&source=images&cd=vfe&ved=0CAIQjRxqFwoTCMDSgOmN2O8CFQAAAAAdAAAAABAD', 'username2');
 INSERT INTO PUBLIC.USER (ID, BIRTH, CAREER_TYPE, GENDER_TYPE, INTRODUCTION, LOCATION_TYPE, PASSWORD, PROFILE_IMAGE, USERNAME) VALUES ('google1', '2021-03-30', 'SENIOR', 'MALE', '자기소개', 'SEOUL', '$2a$10$EUcWG8grzRnfWr0NnGT5ZeKKWq03ALSCgBVKXTqju0xwG1NP.FzW.', 'https://www.google.com/url?sa=i&url=https%3A%2F%2Ffirpeng.tistory.com%2F103&psig=AOvVaw2eQNj_SndJ_UcO5fS-SsSk&ust=1617196831404000&source=images&cd=vfe&ved=0CAIQjRxqFwoTCMDSgOmN2O8CFQAAAAAdAAAAABAD', 'username3');
 
-INSERT INTO PUBLIC.USER_AUTHORITY_MAPPING (ID, AUTHORITY_ID, USER_ID) VALUES (1, 1, 'kakao1');
-INSERT INTO PUBLIC.USER_AUTHORITY_MAPPING (ID, AUTHORITY_ID, USER_ID) VALUES (2, 1, 'google1');
-INSERT INTO PUBLIC.USER_AUTHORITY_MAPPING (ID, AUTHORITY_ID, USER_ID) VALUES (3, 2, 'google1');
+INSERT INTO PUBLIC.USER_AUTHORITY (ID, AUTHORITY_ID, USER_ID) VALUES (1, 1, 'kakao1');
+INSERT INTO PUBLIC.USER_AUTHORITY (ID, AUTHORITY_ID, USER_ID) VALUES (2, 1, 'google1');
+INSERT INTO PUBLIC.USER_AUTHORITY (ID, AUTHORITY_ID, USER_ID) VALUES (3, 2, 'google1');
